theme and util package is moved to :ui:common module
