+++ ui/common/src/main/java/dev/mkeeda/spaceship/ui/common/theme/Color.kt
@@ -1,4 +1,4 @@
-package dev.mkeeda.spaceship.ui.theme
+package dev.mkeeda.spaceship.ui.common.theme
 
 import androidx.compose.ui.graphics.Color
