+++ ui/common/src/main/java/dev/mkeeda/spaceship/ui/common/util/Preview.kt
@@ -1,8 +1,8 @@
-package dev.mkeeda.spaceship.ui.util
+package dev.mkeeda.spaceship.ui.common.util
 
 import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
-import dev.mkeeda.spaceship.ui.theme.SpaceshipTheme
+import dev.mkeeda.spaceship.ui.common.theme.SpaceshipTheme
 
 @Composable
 fun PreviewBackground(content: @Composable () -> Unit) {
