+++ ui/timeline/src/main/java/dev/mkeeda/spaceship/ui/timeline/SenderIcon.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import dev.mkeeda.spaceship.ui.theme.Gray400
+import dev.mkeeda.spaceship.ui.common.theme.Gray400
 
 @Composable
 fun SenderIcon(modifier: Modifier = Modifier) {
